Fixed active quests cache's concurrency issue.
