+++ src/main/java/dev/hugog/minecraft/wonderquests/commands/concrete/CreateRequirementCommand.java
@@ -1,6 +1,6 @@
 package dev.hugog.minecraft.wonderquests.commands.concrete;
 
-import dev.hugog.minecraft.wonderquests.actions.concrete.CreateQuestRequirementAction;
+import dev.hugog.minecraft.wonderquests.actions.implementation.CreateQuestRequirementAction;
 import dev.hugog.minecraft.wonderquests.commands.AbstractPluginCommand;
 import dev.hugog.minecraft.wonderquests.commands.CommandDependencies;
 import dev.hugog.minecraft.wonderquests.injection.factories.ActionsFactory;
