+++ src/main/java/dev/hugog/minecraft/wonderquests/commands/concrete/CreateRewardCommand.java
@@ -1,7 +1,6 @@
 package dev.hugog.minecraft.wonderquests.commands.concrete;
 
-import dev.hugog.minecraft.wonderquests.actions.concrete.CreateQuestRequirementAction;
-import dev.hugog.minecraft.wonderquests.actions.concrete.CreateQuestRewardAction;
+import dev.hugog.minecraft.wonderquests.actions.implementation.CreateQuestRewardAction;
 import dev.hugog.minecraft.wonderquests.commands.AbstractPluginCommand;
 import dev.hugog.minecraft.wonderquests.commands.CommandDependencies;
 import dev.hugog.minecraft.wonderquests.injection.factories.ActionsFactory;
