+++ src/main/java/dev/hugog/minecraft/wonderquests/data/services/QuestsService.java
@@ -1,13 +1,11 @@
 package dev.hugog.minecraft.wonderquests.data.services;
 
 import com.google.inject.Inject;
-import dev.hugog.minecraft.wonderquests.cache.QuestsCache;
-import dev.hugog.minecraft.wonderquests.data.dtos.ActiveQuestDto;
+import dev.hugog.minecraft.wonderquests.cache.implementation.QuestsCache;
 import dev.hugog.minecraft.wonderquests.data.dtos.QuestDto;
 import dev.hugog.minecraft.wonderquests.data.dtos.QuestObjectiveDto;
 import dev.hugog.minecraft.wonderquests.data.dtos.QuestRewardDto;
 import dev.hugog.minecraft.wonderquests.data.dtos.requirements.QuestRequirementDto;
-import dev.hugog.minecraft.wonderquests.data.models.ActiveQuestModel;
 import dev.hugog.minecraft.wonderquests.data.models.QuestModel;
 import dev.hugog.minecraft.wonderquests.data.models.QuestObjectiveModel;
 import dev.hugog.minecraft.wonderquests.data.models.QuestRequirementModel;
